chore: melhora acao de add new list
foi realizado uma modificação que antes impossibilitava de criar lista quando não tivesse nenhuma lista registrada no board, agora é possível adicionar lista mesmo não tendo nenhuma lista no board
