+++ projects/generic-trello/batman/app/src/components/Home/index.jsx
@@ -10,7 +10,7 @@ function page() {
   return (
     <BrowserRouter>
       <Switch>
-        <Route exact path='/login'>
+        <Route exact path='/'>
           <Login />
         </Route>
