+++ projects/generic-trello/batman/app/src/components/List/index.js
@@ -7,7 +7,6 @@ import Dialog from '@material-ui/core/Dialog'
 import FormList from '../Form-List'
 import axios from 'axios'
 import { getToken } from '../../services/auth'
-import Modal from '@material-ui/core/Modal';
 
 export default function List({ data, idList: listIndex }) {
 
@@ -27,37 +26,18 @@ export default function List({ data, idList: listIndex }) {
   }, [urlCards, setCards])
 
   const [open, setOpen] = React.useState(false)
-  const [openList, setOpenList] = useState(false)
 
-  const handleClickOpenList = () => {
-    setOpenList(true)
-  }
-
-  const handleCloseList = () => {
-    setOpenList(false)
+  const handleClose = () => {
+    setOpen(false)
   }
 
   const handleClickOpen = () => {
     setOpen(true)
   }
 
-  const handleClose = () => {
-    setOpen(false)
-  }
 
   return (
     <>
-      <ButtonNew type="button" onClick={handleClickOpenList}>
-        <MdAdd size={24} color='#fff' />
-      </ButtonNew>
-      <Modal
-        open={openList}
-        onClose={handleCloseList}
-        aria-labelledby="simple-modal-title"
-        aria-describedby="simple-modal-description"
-      >
-        <FormList open={openList} idBoard={data.idBoard} handleClose={handleClose} />
-      </Modal>
       <Container done={true}>
 
         <Header>
